+++ hutool-core/src/test/java/cn/hutool/core/util/XmlUtilTest.java
@@ -4,6 +4,7 @@ import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.io.resource.ResourceUtil;
 import cn.hutool.core.map.MapBuilder;
 import cn.hutool.core.map.MapUtil;
+import lombok.Data;
 import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
@@ -173,5 +174,34 @@ public class XmlUtilTest {
 				"//soap:Envelope/soap:Body/ns2:testResponse/return",
 				document,XPathConstants.STRING);//
 		Assert.assertEquals("2020/04/15 21:01:21", value);
+	}
+
+	@Test
+	public void xmlToBeanTest(){
+		final TestBean testBean = new TestBean();
+		testBean.setReqCode("1111");
+		testBean.setAccountName("账户名称");
+		testBean.setOperator("cz");
+		testBean.setProjectCode("123");
+		testBean.setBankCode("00001");
+
+		final Document doc = XmlUtil.beanToXml(testBean);
+		Assert.assertEquals(TestBean.class.getSimpleName(), doc.getDocumentElement().getTagName());
+
+		final TestBean testBean2 = XmlUtil.xmlToBean(doc, TestBean.class);
+		Assert.assertEquals(testBean.getReqCode(), testBean2.getReqCode());
+		Assert.assertEquals(testBean.getAccountName(), testBean2.getAccountName());
+		Assert.assertEquals(testBean.getOperator(), testBean2.getOperator());
+		Assert.assertEquals(testBean.getProjectCode(), testBean2.getProjectCode());
+		Assert.assertEquals(testBean.getBankCode(), testBean2.getBankCode());
+	}
+
+	@Data
+	public static class TestBean{
+		private String ReqCode;
+		private String AccountName;
+		private String Operator;
+		private String ProjectCode;
+		private String BankCode;
 	}
 }